# event_go/internal/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulmarker Event API",
    "version": "1.0.0",
    "description": "Versioned API contract for authentication, organization authorization, organizers, events, registrations, admissions, checkins, discussions and tickets. Legacy /api routes remain available for one compatibility cycle."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "registerUser",
        "tags": ["Auth"],
        "summary": "Register a user and issue a JWT",
        "requestBody": { "$ref": "#/components/requestBodies/RegisterUserRequest" },
        "responses": {
          "201": { "$ref": "#/components/responses/Created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "loginUser",
        "tags": ["Auth"],
        "summary": "Authenticate and issue a JWT",
        "requestBody": { "$ref": "#/components/requestBodies/LoginRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logoutUser",
        "tags": ["Auth"],
        "summary": "Revoke all current user sessions",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "operationId": "requestPasswordReset",
        "tags": ["Auth"],
        "summary": "Request an account-enumeration-safe password reset email",
        "requestBody": { "$ref": "#/components/requestBodies/PasswordResetRequest" },
        "responses": {
          "202": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "operationId": "confirmPasswordReset",
        "tags": ["Auth"],
        "summary": "Consume a one-time password reset token",
        "requestBody": { "$ref": "#/components/requestBodies/PasswordResetConfirmRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/auth/recovery-email/confirm": {
      "post": {
        "operationId": "confirmRecoveryEmail",
        "tags": ["Auth"],
        "summary": "Confirm a one-time recovery email verification token and revoke old sessions",
        "requestBody": { "$ref": "#/components/requestBodies/RecoveryEmailConfirmRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/me/registrations": {
      "get": {
        "operationId": "listMyRegistrations",
        "tags": ["Me"],
        "summary": "List registrations owned by the current user",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/me/admissions": {
      "get": {
        "operationId": "listMyAdmissions",
        "tags": ["Me", "Admissions"],
        "summary": "List current and revoked free-event admission credentials",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/me/activities": {
      "get": {
        "operationId": "listMyActivities",
        "tags": ["Me", "Admissions"],
        "summary": "List a stable unified timeline of registrations and admissions",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/me/recovery-email/request": {
      "post": {
        "operationId": "requestRecoveryEmail",
        "tags": ["Me", "Auth"],
        "summary": "Request a password-confirmed recovery email verification link",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/RecoveryEmailRequest" },
        "responses": {
          "202": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/me/notifications": {
      "get": {
        "operationId": "listNotifications",
        "tags": ["Me", "Notifications"],
        "summary": "List notifications owned by the current user",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "name": "unread_only", "in": "query", "schema": { "type": "boolean", "default": false } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/me/notifications/unread-count": {
      "get": {
        "operationId": "getNotificationUnreadCount",
        "tags": ["Me", "Notifications"],
        "summary": "Count unread notifications owned by the current user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/me/notifications/{notificationId}/read": {
      "parameters": [{ "$ref": "#/components/parameters/NotificationID" }],
      "put": {
        "operationId": "markNotificationRead",
        "tags": ["Me", "Notifications"],
        "summary": "Mark one current-user notification as read",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/me/notifications/read-all": {
      "put": {
        "operationId": "markAllNotificationsRead",
        "tags": ["Me", "Notifications"],
        "summary": "Mark all current-user notifications as read",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/me/organizations": {
      "get": {
        "operationId": "listMyOrganizations",
        "tags": ["Me", "Organizations"],
        "summary": "List the current user's active memberships and effective tenant capabilities",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/organizations": {
      "post": {
        "operationId": "createOrganization",
        "tags": ["Organizations"],
        "summary": "Create an organization workspace owned by the current user",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateOrganizationRequest" },
        "responses": {
          "201": { "$ref": "#/components/responses/Created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/organization-invitations/accept": {
      "post": {
        "operationId": "acceptOrganizationInvitation",
        "tags": ["Organizations"],
        "summary": "Accept an organization invitation as the current user",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/AcceptOrganizationInvitationRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/organizations/{organizationId}/session": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "getOrganizationSession",
        "tags": ["Organizations"],
        "summary": "Resolve a live tenant principal and its effective capabilities",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/organizations/{organizationId}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "getOrganizationWorkspace",
        "tags": ["Organizations"],
        "summary": "Get the current tenant workspace and organizer profile",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/organizations/{organizationId}/members": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "listOrganizationMembers",
        "tags": ["Organizations"],
        "summary": "List members visible to the current tenant principal",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/organizations/{organizationId}/members/{memberId}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }, { "$ref": "#/components/parameters/OrganizationMemberID" }],
      "put": {
        "operationId": "updateOrganizationMember",
        "tags": ["Organizations"],
        "summary": "Change an organization member role within tenant policy",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/UpdateOrganizationMemberRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "revokeOrganizationMember",
        "tags": ["Organizations"],
        "summary": "Revoke an organization member within tenant policy",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/organizations/{organizationId}/owner-transfer": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "post": {
        "operationId": "transferOrganizationOwnership",
        "tags": ["Organizations"],
        "summary": "Transfer organization ownership to another active member",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/TransferOrganizationOwnerRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/organizations/{organizationId}/invitations": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "listOrganizationInvitations",
        "tags": ["Organizations"],
        "summary": "List organization invitations without secret token material",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "createOrganizationInvitation",
        "tags": ["Organizations"],
        "summary": "Invite a user to the organization and deliver a one-time link",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateOrganizationInvitationRequest" },
        "responses": {
          "201": { "$ref": "#/components/responses/Created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/organizations/{organizationId}/invitations/{invitationId}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }, { "$ref": "#/components/parameters/OrganizationInvitationID" }],
      "delete": {
        "operationId": "revokeOrganizationInvitation",
        "tags": ["Organizations"],
        "summary": "Revoke a pending organization invitation",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/organizations/{organizationId}/audits": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "listOrganizationAudits",
        "tags": ["Organizations", "Audit"],
        "summary": "List tenant-scoped organization audit logs",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/organizations/{organizationId}/events": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "post": {
        "operationId": "createOrganizationEvent", "tags": ["Organizations", "Events"],
        "summary": "Create an event inside the authorized tenant scope", "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateEventRequest" },
        "responses": { "201": { "$ref": "#/components/responses/Created" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "get": {
        "operationId": "listOrganizationEvents", "tags": ["Organizations", "Events"],
        "summary": "List only events owned by the authorized tenant", "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }],
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/organizations/{organizationId}/events/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }, { "$ref": "#/components/parameters/EventID" }],
      "get": {
        "operationId": "getOrganizationEvent", "tags": ["Organizations", "Events"], "summary": "Get a tenant-scoped event", "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "put": {
        "operationId": "updateOrganizationEvent", "tags": ["Organizations", "Events"], "summary": "Update a tenant-scoped event", "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/UpdateEventRequest" },
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "operationId": "deleteOrganizationEvent", "tags": ["Organizations", "Events"], "summary": "Delete a tenant-scoped event", "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
...
    "/organizations/{organizationId}/events/{id}/registrations": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }, { "$ref": "#/components/parameters/EventID" }],
      "get": {
        "operationId": "listOrganizationEventRegistrations", "tags": ["Organizations", "Registrations"], "summary": "List registrations for a tenant-scoped event", "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }],
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/organizations/{organizationId}/events/{id}/registrations/export": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }, { "$ref": "#/components/parameters/EventID" }],
      "get": {
        "operationId": "exportOrganizationEventRegistrations", "tags": ["Organizations", "Registrations"], "summary": "Export a tenant-scoped registration CSV", "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "UTF-8 CSV export" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "AdminToken": { "type": "apiKey", "in": "header", "name": "X-Admin-Token", "description": "Platform administrator credential; it does not grant an organization membership or tenant capability." }
    },
...
    "requestBodies": {
      "RegisterUserRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterUserRequest" } } } },
      "LoginRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
...
    },
    "responses": {
      "OK": { "description": "Successful response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/APIResponse" } } } },
      "Created": { "description": "Resource created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/APIResponse" } } } },
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
      "Conflict": { "description": "Business state conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
...
  }
}